backend/python/requirements.txt
fastapi
uvicorn
gemini
pydantic
google.generativeai
